fix user count, add roles to list, add create subcommand

+++ src/php/wp-cli/commands/internals/user.php
@@ -18,11 +18,12 @@ class UserCommand extends WP_CLI_Command {
   public function all( $args, $assoc_args ) {
     global $blog_id;
 
-    $users = get_users("blog_id=$blog_id");
-    $fields = array('ID', 'user_login', 'display_name', 'user_email',
-      'user_registered', 'user_status');
     $table = new \cli\Table();
-    $table->setHeaders($fields);
+    $users = get_users("blog_id=$blog_id&fields=all_with_meta");
+    $fields = array('ID', 'user_login', 'display_name', 'user_email',
+      'user_registered');
+
+    $table->setHeaders( array_merge($fields, array('roles')) );
 
     foreach ( $users as $user ) {
       $line = array();
@@ -30,14 +31,14 @@ class UserCommand extends WP_CLI_Command {
       foreach ( $fields as $field ) {
         $line[] = $user->$field;
       }
+      $line[] = implode( ',', $user->roles );
 
       $table->addRow($line);
     }
 
     $table->display();
 
-    $total = count_users();
-    WP_CLI::line( 'Total: ' . $total['total_users'] . ' users' );
+    WP_CLI::line( 'Total: ' . count($users) . ' users' );
   }
 
 	/**
@@ -46,12 +47,20 @@ class UserCommand extends WP_CLI_Command {
 	 * @param array $args
 	 * @param array $assoc_args
 	 **/
-	public function users( $args, $assoc_args ) {
+	public function create( $args, $assoc_args ) {
 		global $blog_id;
 
+    $user_login = $args[0];
+    $user_email = $args[1];
+
+    if ( ! $user_login || ! $user_email ) {
+      WP_CLI::error("Login and email required (see 'wp user help').");
+    }
+
 		$defaults = array(
-			'count' => 100,
 			'role' => get_option('default_role'),
+      'user_pass' => wp_generate_password(),
+      'user_registered' => strftime( "%F %T", time() ),
 		);
 
 		extract( wp_parse_args( $assoc_args, $defaults ), EXTR_SKIP );
@@ -63,29 +72,25 @@ class UserCommand extends WP_CLI_Command {
 			exit;
 		}
 
-		$user_count = count_users();
+    $user_id = wp_insert_user( array(
+      'user_email' => $user_email,
+      'user_login' => $user_login,
+      'user_pass' => $user_pass,
+      'user_registered' => $user_registered,
+      'display_name' => $display_name,
+      'role' => $role,
+    ) );
 
-		$total = $user_count['total_users'];
+    if ( is_wp_error($user_id) ) {
+      WP_CLI::error( $user_id->get_error_message() );
+    } else {
+      if ( false === $role ) {
+        delete_user_option( $user_id, 'capabilities' );
+        delete_user_option( $user_id, 'user_level' );
+      }
+    }
 
-		$limit = $count + $total;
-
-		for ( $i = $total; $i < $limit; $i++ ) {
-			$login = sprintf( 'user_%d_%d', $blog_id, $i );
-			$name = "User $i";
-
-			$user_id = wp_insert_user( array(
-				'user_login' => $login,
-				'user_pass' => $login,
-				'nickname' => $name,
-				'display_name' => $name,
-				'role' => $role
-			) );
-
-			if ( false === $role ) {
-				delete_user_option( $user_id, 'capabilities' );
-				delete_user_option( $user_id, 'user_level' );
-			}
-		}
+    WP_CLI::line( "Created user $user_id" );
 	}
 
 	/**
@@ -93,8 +98,8 @@ class UserCommand extends WP_CLI_Command {
 	 */
 	public static function help() {
 		WP_CLI::line( <<<EOB
-usage: wp generate posts [--count=100] [--type=post] [--status=publish]
-   or: wp generate users [--count=100] [--role=<role>]
+usage: wp user all
+   or: wp user create <user_login> <user_email> [--role=<default_role>]
 EOB
 	);
 	}
